Add files via upload
The Uncontested, True, What Some May Say The Greatest School Ever In Existence, A School No One Can Match. Watch Us As We Reach For The STARS And The Ever BEYOND. The Absolute True Monarch, Tyrant Of The World, An Absolute Power, One That Shall Be Unrivaled for Aeons To Come. Watch Us Rule The Whole World, And Annihilate All Those Opposing Us, As Our True Perfection Shine Upon This Sullen Dark World, Witness As We Restore The Dreams Of The Billions That Came Before Us And Those Who Shall Eventually Come